// JavaScriptProjects/ReactProjects/warehouse_wars/front-end/ww-react-ts/src/types/game/ClientSend.ts
export interface RequestName extends GamePackage {
    type: "request_name";
    id?: string;
    name: string;
}